Moved login back to new iqsign folder, changed routes accordingly.
